Make the SWIFT output plugin default to double precision

+++ src/plugins/output_swift.cc
@@ -369,7 +369,7 @@ public:
 
 namespace
 {
-output_plugin_creator_concrete<swift_output_plugin<float>> creator1("SWIFT");
+output_plugin_creator_concrete<swift_output_plugin<double>> creator1("SWIFT");
 } // namespace
 
 #endif
